+++ CHANGELOG.md
@@ -14,12 +14,13 @@
 # Changelog of EIRENE repository:
 ---
 
-## X. Bonnin - SOLPS_coupling - 72b40d313154040c452b29696b76a0ce3c22a91d
+## X. Bonnin - SOLPS_coupling - 6b15f6460fc69c156f330dfdde9b35d138b9e20f
 
 ### Based on SOLPS_bug_fixes - 765b12f7f7ab978e8340d215d330ec49c6c5417b
 
 ** Changes **
 
+- New IMPROVED MPI parallelization strategy
 - Bug fixes for time-dependent runs and ensuring we always have a filled-in background for coupled runs
 - Debugged hybrid OpenMP+MPI parallelization
 - Added passing of ZI to SAMSRF